Implement `MysqlConnection::execute_returning_count`
This adds the initial implementation of prepared statement support for
MySQL. This does not handle prepared statements which return values, nor
does it implement any form of prepared statement caching.

Prepared statements in MySQL operate on a structure called `MYSQL_BIND`
both for input and output. This thing is a fucking nightmare to work
with. We need to give it the structs in an array, which is behind a
pointer. We are responsible for allocating, managing, keeping alive, and
cleaning up the `MYSQL_BIND` structures. Additionally, it has several
fields which are also behind pointers, and we are also responsible for
keeping those values alive. Looking through their API docs, I see no
reasonable cause to have any of the fields except the buffer behind a
pointer. Still, that's the API we have to work with, so we have a bunch
of random vecs on the struct that only really exist so that we keep
these pointers alive.